[turboshaft][wasm] Load the stack check limit address only once
The previous implementation of StackCheck lowering required 2 LoadOp
per stack check: one for the limit address, and one for the limit
itself. With this CL, the limit address is only loaded once at the
begining.

While it's "just one less load", this improves performance of some
benchmarks by up to 5-10%.

Note that the StackCheck lowering always unconditionally loads the
stack limit once. Since this is done in the last phase, we can rely
on OptimizationPhase to remove it if it has no uses. So,
RecreateSchedule and InstructionSelector now also ignore Turboshaft
operations that have no uses.

Bug: v8:14108
Change-Id: Iff994bebe1dd3579987ec883fd55178f418785c1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4903651
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90240}