[sandbox] Allocate trusted space outside of the sandbox
With this CL, we now use a special page allocator to allocate the
backing memory for trusted space pages. This allocators allocates memory
out of the new TrustedRange, which is a virtual address space
initialized once per process and located outside the sandbox.
We also introduce a IsTrustedSpaceObject predicate that can be used to
determine whether an object lives in trusted space (and so outside of
the main pointer compression cage).

Bug: chromium:1473677
Change-Id: Ie6468550a16dcc159ac2f729e1f4c08f7846c2ef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4838714
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90260}