[csa] Truncate with TruncateFloat64To... instead of ChangeFloat64To...
In CodeStubAssembler::TryToInptr, we used the pattern
  x = ChangeFloat64ToIntPtr(value);
  y = RoundIntPtrToFloat64(x);
  x == y
to decide if value is an integer. However, Change-operators assume
that they preserve the value (that the conversion is reversible).
Turboshaft exploits this fact and optimizes the above sequence into
  x == x
which is clearly incorrect. This is observable with the new
Turboshaft builtin pipeline (--turboshaft-csa).
This is fixed by using a proper truncation operation.

Change-Id: I38571bf5f06b956e2d48f9431549925ff0b89651
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4913735
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90264}