[debug] Fix crash when pausing while entering a 'with' scope
This CL fixes a bug where we get the "paused scope" wrong. When
"pause on exception" is enabled we can pause on the early bytecodes
of a 'with' scope. The example here is attempting to load a undefined
variable from the context which results in a ReferenceError.

The mapped source position of the load bytecode points directly at the
beginning of the scope, which we normally exclude. The fix is to also
include the start scope boundary for 'with' scopes, similar to
'class' scopes.

Fixed: chromium:1482394
Change-Id: I1b2051c2dab6533d38ca69284710ffc5d54b4496
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4916656
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90273}