[heap] Move LAB creation out of space construction
LABs for the main thread and GC evacuation threads were always
created in the space constructor. However, this is problematic as
we cannot use any virtual methods on the space while setting up the
LAB. Our goal is anyways to move the LAB out of the space entirely,
this CL is a step towards that goal.

LABs are still stored in SpaceWithLinearArea but are now created
using SpaceWithLinearArea::CreateMainAllocator() in the
HeapAllocator and the EvacuationAllocator class.

Bug: chromium:1480975
Change-Id: Icdac1d9f0e2d52e18e17e322eebdb54dd14ce57e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4916171
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90274}