[liftoff] Rename StackTransferRecipe to ParallelMove
The StackTransferRecipe is used for multiple things now, not only to
transfer one value stack to another one (in MergeStackWith and friends;
this is where the original name came from).

It's also 340 lines of implementation by now, so move it to a separate
header file.

This creates the situation that we now have headers (parallel-move.h and
liftoff-assembler-<arch>.h) that want to include either other.
We thus apply V8's default solution and split off an "inl header" for
the two where we define some of the inlined functions that are declared
in the actual header.
To that end, we also rename the existing liftoff-assembler-<arch>.h
header to liftoff-assembler-<arch>-inl.h, because those headers also
just define functions that have been declared in the actual header. This
also make us obey the rule that non-inl-headers cannot include
inl-headers.

R=dlehmann@chromium.org

Bug: v8:14101
Change-Id: Ie5f6d3a21f4986143855a83826d00d706faa4b7e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4905773
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90303}