[heap] Remove SpaceWithLinearArea::main_allocator() getter
While we cannot remove the MainAllocator* in SpaceWithLinearArea at
the moment, we can already remove the getter for it.

Callers need to get to the MainAllocator instances through the
HeapAllocator class instead.

Drive-by: Rename PublishPendingAllocations() to
PublishMainThreadPendingAllocations() to emphasize that we only
publish pending allocation for the main thread.

Bug: chromium:1480975
Change-Id: I5d0416694db580264c474b45bd19a6188869b8f8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4927168
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90374}