Add name hash mask to quickly check for no integer/forwarding-index
Add a new mask to check that a hash doesn't contain an integer index or
a forwarding index.
This allows for quick checks that a hash is guaranteed to not contain an
integer index, as long as false negatives (i.e. a false value doesn't
mean that the hash does represent an integer index) are allowed.

Bug: chromium:1492013
Change-Id: Id6bfaaa14d2403c659dd7fdaad790cf2aaffe94e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4935110
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90404}