[wasm] Use C-call for atomic.notify
There is no need to call a runtime function via the CEntry stub. The
method does not allocate or otherwise trigger GC, and its only arguments
are the wait location and the number of workers to wake.

Thus this CL implements this via a direct C-call.

R=dlehmann@chromium.org, marja@chromium.org

Bug: v8:14101
Change-Id: I5063dc7c5b82b6efea60811bf5155adaa498edcf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4898689
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90392}