[BS] Do not touch all bytes in debug mode
The `DebugCheckZero` check (which is only enabled for Wasm anyway)
ensures that a newly allocated backing store is correctly
zero-initialized.
This causes a significant and unnecessary increase in the amount of
allocated memory in debug builds and causes infra failures.

Thus skip this test completely. We have existing tests that check that
freshly allocated or grown memory is zero-initialized.

R=ahaas@chromium.org
CC=machenbach@chromium.org

Bug: v8:14373, v8:4153
Change-Id: I9ecb39fcdeb4e94d7787962907067266f44de44f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4935418
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90396}