[sandbox] Ensure EPT evacuation entries are always resolved during GC
After crrev.com/c/4898680 we may now bail out from
ResolveEvacuationEntryDuringSweeping if the external pointer slot has
been deinitialized in the meantime. However, in that case we must clear
the evacuation entry (for example by adding it to the freelist) as we
would otherwise try to process it again during the next GC, by which
time the address stored in the evacuation entry may no longer be valid
as the underlying object may have been freed or moved.

Bug: chromium:1492019
Change-Id: I8b1484a6645668052c310e0dae79141b5f4ee17c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4938835
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90420}