[heap] Remove UpdateInlineAllocationLimit() methods
Adding allocation observers may decrease the number of bytes to the
next allocation step. This means that we need to shrink the current
LABs in order to run the allocation observer when required. This
logic is implemented by UpdateInlineAllocationLimit().

However, if there exists no LAB this method becomes a no-op. This CL
ensures that at any time we add an allocation observer to
MainAllocator, there is no active LAB. With this invariant we
can get rid of UpdateInlineAllocationLimit().

In order to achieve this in production code, the CL only needs to free
all LABs at a GC. GCs are infrequent enough that this shouldn't hurt
performance.

The other downside of this CL is, that allocation observers can't
add allocation observers in an allocation observer step anymore.
However, we currently don't use this and if this use case ever comes
up, we can likely implement this logic with a single allocation
observer and some state as well.

Allocation observer can still remove themselves in an allocation
step. This is something we actually use and removing an allocation
observer can only increase step size. So it should be safe even
without adjusting LAB sizes.

Bug: chromium:1480975
Change-Id: I41489ef70ed8c63808d53885ec9a458e67fd0f9f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4938832
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90455}