Refactor argument passing for templated messages
Pass a vector of arguments instead of a fixed number, where some are
either null handles or undefined, and in the end we use "undefined" for
missing arguments and ignore any additional ones.

This will allow us to check that the number of provided arguments
matches what the template string expects.

For now, we have an allowlist of templates where we currently do not
pass the expected number of arguments. Follow-up CLs will successively
shrink that list until the fallback code can finally be removed.

R=jkummerow@chromium.org

Bug: v8:14386
Cq-Include-Trybots: luci.chromium.try:linux-rel
Change-Id: I4b23c09b9c3dd28e53cce882eb5c5367bcdc81f8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4946850
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90472}