[test] Add process-memory probe and track consumption
This adds a probing mechanism for tests executed on desktop platforms.
The probe tracks memory statistics with psutil. The results are
added to the resultsdb entry and the json test results.
For the json test results, only the top 20 tests are tracked,
similar to duration.

When running locally or when psutil is not available, the probe will
just stay inactive. There is no significant overhead when probing
every 0.2 seconds. Very fast tests will show 0 results.

Similar to test duration, we can use the information from the json
to list the top memory-hungry tests, which are not yet marked
as HEAVY in the status file.

Such a listing will be implemented as a follow up on the infra side.

Bug: v8:14373
Change-Id: Ief162d5ef88bdd7792a953505feeb5a8f1e4fe94
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4946490
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@google.com>
Cr-Commit-Position: refs/heads/main@{#90492}