[wasm][liftoff] Use root register for stack checks
With this CL, Liftoff loads the stack limit with the root register and
not from the WebAssembly instance. The offset of the stack limit from
the isolate root is added to IsolateData, since Liftoff does not have
access to the isolate to reuse the way how JavaScript calculates the
offset.

R=clemensb@chromium.org, jgruber@chromium.org

Bug: v8:14393
Change-Id: I89e80ba8b38f14fef685fcd68ebeee368d9c8e29
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4952153
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90494}