Propagate ContinuationPreservedEmbedderData through thenables
Because the `then` function of a thenable is run within a microtask for PromiseResolveThenableJobTask, the state of the stored ContinuationPreservedEmbedderData may not match what is expected within that `then` call unless the microtask which makes that call also propagates the value as is done with PromiseFulfillReactionJobTask and PromiseRejectReactionJobTask.

This change simply copies what is already done with PromiseFulfillReactionJobTask and PromiseRejectReactionJobTask to copy the ContinuationPreservedEmbedderData state on creation of the job tasks and restoring the state while running.

  setContinuationPreservedEmbedderData('data')
  await Promise.resolve({
    then (fulfill) {
      // Without this change the state here will be empty
      console.log(getContinuationPreservedEmbedderData())
      setImmediate(fulfill)
    }
  })

Change-Id: Iebc568deb7dcc43dfeed0bf5a0d140df8614f9e8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4674242
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90668}