[object] Make the non-JS oddball markers into holes
All the marker oddballs (uninitialized value, optimized out, etc.),
fulfil a role closer to "hole" than to "oddball", in that they are not
expected to escape into user code. Most places that check for oddballs
are only really checking for the user-visible oddballs (null, undefined,
true and false), and it's both more elegant and more secure to be
explicit about caring about JS oddballs vs. internal markers.

This requires a few typer test changes, since the uninitialized value is
now a hole and therefore not a singleton as far as the typer is
concerned.

Change-Id: I191fc89b3a2e6407e6cb2156ac0d7c24b7947894
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4952095
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90516}