[wasm][turboshaft] Load stack limit from the root register
With this CL, Turboshaft uses the root register to load the stack limit
in WebAssembly. The JavaScript stack check code gets slightly adjusted
for code sharing with WebAssembly, the resulting machine code should be
the same. The existing code relied on pointer arithmetic in the macro
assembler to calculate the correct offset from the root register, the
code in this CL embeds the correct offset directly into the Turboshaft
graph.

R=nicohartmann@chromium.org

Bug: v8:14393
Change-Id: Ib7d4c65935e1b0adfbce207cea675e9cc75c97d5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4962098
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90555}