Reland "[wasm][turboshaft] Loop unrolling"
The original CL got reverted because Phis that resulted in
LoadRootRegisterOps reached the instruction selector. Normally a
LoadRootRegisterOp does not have a use in the instruction selector, and
VisitLoadRootRegister does not get called. However, in VisitPhi all
inputs get used, and therefore also the LoadRootRegisterOp.

The problem could be solved on several levels:
1) Remove the UNREACHABLE in VisitLoadRootRegister and replace it with a
   nop;
2) Eliminate recursive structures of Phis whose input is always either a
   Phi from the same structure or a fixed value.
3) Never create a Phi which has a LoadRootRegister as an input.

This CL implements solution 3. Solution 1 may hide errors, solution 2
may be a bit too complex for just this single problem. Solution 2 may
eventually be worth it, but so far it only introduced compilation
overhead and did not result in any benefits.

Original change's description:
> [wasm][turboshaft] Loop unrolling
>
> This CL enables the Turboshaft loop unrolling implementation for
> WebAssembly. For WebAssembly, slightly higher limits seem to be useful,
> and a different stack check is used. Otherwise the loop unrolling is
> unchanged.

Bug: v8:14108
Change-Id: I5fe9c5370e815b0c5d2f8bbcc634cb506da3971e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4966484
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90554}