[test] Use correct PID for logging memory stats on Linux
On Linux we start subprocesses with shell=True, stemming from
https://crbug.com/v8/8889 (reversing that change might be possible,
but risky, according to the bug).

The shell settings starts an extra shell wrapper that wraps the
V8 test process. On Debian this is optimized using exec, leading to
just one PID, while on Ubuntu (e.g. our bots) this creates a child
PID.

This change retrieves the child PID if possible, in order to use
that one for memory probing.

Bug: v8:14373
Change-Id: Ic7fd9df93de08959477ba5c99b269771076a7968
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4968342
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@google.com>
Cr-Commit-Position: refs/heads/main@{#90567}