[wasm] Avoid temporary heap allocations in module decoding
Function signatures will eventually be stored in zone memory, but the
current implementation allocates two temporary std::vectors for
parameters and returns.
This CL uses a SmallVector for parameters, and allocates the final zone
memory as soon as we decoded the return count. This avoids copying the
return types, and avoids dynamic memory allocation and deallocation for
the parameters in the majority of cases (up to 8 parameters).

R=manoskouk@chromium.org

Bug: v8:14402
Change-Id: Ieaccf95e89f2690a5429cf505a93d0c86d95368f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4971834
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90564}