[wasm][turboshaft] WasmGCTypeReducer: Type call results
While it seems like a simpler change to just add the wasm::ValueType to
the CallOp (it doesn't actually cost any memory on 64 bit platforms due
to alignment), this would unfortunately require many
 #if V8_ENABLE_WEBASSEMBLY
resulting in different signatures for reducers of callops depending
whether webassembly is enabled or not.

So instead the more generic WasmTypeAnnotationOp is added to the graph
to store this information in a generic way. The node is only generated
for reference types (as we don't perform any type optimizations on
non-reference types).

Bug: v8:14108
Change-Id: I65527b080ce5bb55084441c4bb9bfde752d8a333
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4976266
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90618}