[base] Introduce Hasher
We are pretty inconsistent currently in the use of the
base::hash_combine functions. There is a templatized function which
hashes multiple values and combines the computed hashes. The function to
combine two hashes is also called base::hash_combine, so it can happen
that we (or the compiler) accidentally chooses that implementation,
which then does *not* hash the values but just combines them (which
results in poor hashes).

The Hasher also provides an explicit AddHash method which works like
the previous hash_combine(size_t, size_t), i.e. it assumes that the
RHS is a hash and not a plain integer value (which might have less
uniform distribution than a hash).

This CL also adds a hash_value function for each type with a
hash_value member function, which allows to implement hash_value
within a class or struct definition. This is particularly useful if
the class is private (hence it would be difficult to implement a
function that accepts this class as parameter).

R=leszeks@chromium.org

Bug: v8:12791
Change-Id: Ie6b9daf7a957660791e007951310ee6c60a6e2d1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4983675
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90662}