Don't store WasmCode references in CallSiteInfo's code_object field
The code_object field is only used for computing source positions, but
for Wasm frames these are computed independently of the code object. As
such, we don't need to allocate a Managed wrapper object for the
WasmCode. This change is also beneficial for the sandbox: now the
CallSiteInfo's code_object can only ever be Code or BytecodeArray, both
of which can be referenced via indirect pointers. If we also had to
support other objects that live inside the sandbox, then we'd need two
separate fields.

Bug: chromium:1473677
Change-Id: I4488bb169ef5b275fe86a114a94ab80d40e2f6b7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4983676
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90653}