[test] Mark tests with large memory consumption
This reorganizes tests marked as HEAVY in status files. The new data
is from recent memory probing on various platforms, tracked in the
UI.

The previous data was mostly anecdotal (i.e. based on certain
test failures). This old data might have been set incorrectly,
as we might have marked tests as heavy just because they failed,
while the real heavy test was running in the background.

Hence, tests not showing up in the list of tests with large memory
consumption are now unmarked.

Additionally mjsunit/huge-typedarrays is now only run in the default
variant on Windows, as it always ran in the background when
any memory problems appeared in other tests, see analysis in
issue v8:14373.

Bug: v8:14373
Change-Id: I69f3b4a2b63ab00aff3a5c2ed30e437a168037a3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4989322
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90718}