[sandbox] Introduce BytecodeWrapper type
This is a HeapObject that only contains a (trusted pointer) reference to
a BytecodeArray object. However, in contrast to the BytecodeArray
itself, the wrapper object lives inside the sandbox. This allows it to
be used in cases where a BytecodeArray is referenced from a collection
of tagged pointers (e.g. a FixedArray).
This change also implements support for indirect-pointer back-references
in the snapshot serializer as that is required to properly handle the
bytecode <-> wrapper cycle.

Drive-By: Rename kIndirectPointerSlotSize to simply kIndirectPointerSize

Bug: chromium:1472252
Change-Id: I7bf3927c680d234fd4559db4459510e8af81883b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4993365
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#90720}